requirements.txt
torch>=1.8
torchvision>=0.9
mmcv-full>=1.3.0,<2.0.0
mmsegmentation>=0.20.0,<1.0.0
numpy
Pillow
tqdm
